refactored into modules.  Still needs Edit feature and CSS.
