+++ api/nutshell.json
@@ -40,7 +40,7 @@
       "id": 1,
       "taskName": "Chores",
       "dueDate": "2020/05/21",
-      "isCompleted": true,
+      "isCompleted": false,
       "userId": 1
     },
     {
@@ -49,6 +49,139 @@
       "isCompleted": false,
       "userId": 2,
       "id": 2
+    },
+    {
+      "taskName": "Nutshell Project",
+      "dueDate": "2020-05-22",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 3
+    },
+    {
+      "taskName": "Take dog hiking",
+      "dueDate": "2020-05-23",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 4
+    },
+    {
+      "taskName": "Take breaks",
+      "dueDate": "2020-05-19",
+      "isCompleted": true,
+      "userId": "1",
+      "id": 5
+    },
+    {
+      "taskName": "Laundry",
+      "dueDate": "2020-05-20",
+      "isCompleted": false,
+      "userId": "2",
+      "id": 6
+    },
+    {
+      "taskName": "Study",
+      "dueDate": "2020-05-22",
+      "isCompleted": false,
+      "userId": "2",
+      "id": 7
+    },
+    {
+      "taskName": "sleep",
+      "dueDate": "2020-05-19",
+      "isCompleted": true,
+      "userId": "2",
+      "id": 8
+    },
+    {
+      "taskName": "test",
+      "dueDate": "2020-05-19",
+      "isCompleted": true,
+      "userId": "2",
+      "id": 9
+    },
+    {
+      "taskName": "test agdagdagda",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "2",
+      "id": 10
+    },
+    {
+      "taskName": "dagadgdag",
+      "dueDate": "2020-05-19",
+      "isCompleted": true,
+      "userId": "1",
+      "id": 11
+    },
+    {
+      "taskName": "shjfhshfshfshfs",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 12
+    },
+    {
+      "taskName": "aaeeee",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 13
+    },
+    {
+      "taskName": "aaeeee",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 14
+    },
+    {
+      "taskName": "fffff",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 15
+    },
+    {
+      "taskName": "tatadat",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 16
+    },
+    {
+      "taskName": "gsgsgsgeerer",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 17
+    },
+    {
+      "taskName": "afsafafaaf",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 18
+    },
+    {
+      "taskName": "safdafda",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 19
+    },
+    {
+      "taskName": "newest test",
+      "dueDate": "2020-05-19",
+      "isCompleted": false,
+      "userId": "1",
+      "id": 20
+    },
+    {
+      "taskName": "one test",
+      "dueDate": "2020-05-19",
+      "isCompleted": true,
+      "userId": "1",
+      "id": 21
     }
   ],
   "events": [
